Testing the converters

+++ src/java/entitiesConverters/TrackConverter.java
@@ -8,7 +8,6 @@ package entitiesConverters;
 
 import com.mongodb.BasicDBObject;
 import com.mongodb.DBObject;
-import entities.Album;
 import entities.Track;
 
 /**
@@ -28,7 +27,7 @@ public abstract class TrackConverter
     public static Track convertDBObjectToTrack(DBObject dbTrack)
     {
         //TODO: Verify if dbTrack is valid and throw exception otherwise
-        Track track = new Track(Short.parseShort(dbTrack.get("number").toString()), dbTrack.get("name").toString());
+        Track track = new Track(Integer.parseInt(dbTrack.get("number").toString()), dbTrack.get("name").toString());
         return track;
     }
 }
